refactor(web): re-home ItemSummary and ThemedIcon
#934 grouped `components/` into `chrome/`, `summaries/`, and `ui/`,
leaving no files flat at the root. This branch forked before that, so
merging develop would have dropped the two new components back at the
root and made the grouping partial again.

`ItemSummary` only ever composes character and weapon summaries, so it
joins them in `summaries/` and they reach it by relative import, matching
how `features/teams/` files reference their siblings. `ThemedIcon` is a
presentational primitive with no domain knowledge — the header logo and
both collection filter bars are its eventual callers, not just the
summaries — so it sits in `ui/` alongside the other primitives.
